+++ HISTORY.rst
@@ -2,6 +2,11 @@
 History
 =======
 
+---
+
+- Fix: dict data type on update.
+- Fix: Use internal remove method to delete old node.
+
 0.1.3
 ------------------